pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "testagent"
version = "0.1.0"
description = "AI-powered testing agent platform for App/Web/API"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "TestAgent Team" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "typer>=0.9",
    "rich>=13.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "httpx>=0.27",
    "keyring>=25.0",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "aiosqlite>=0.19",
    "psycopg[binary]>=3.1",
    "asyncpg>=0.29",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "mcp>=1.0",
    "chromadb>=0.4",
    "meilisearch-python-async>=1.5",
    "openai>=1.10",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "pytest-timeout>=2.2",
    "ruff>=0.3",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
]
milvus = [
    "pymilvus>=2.4",
]

[project.scripts]
testagent = "testagent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["testagent"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "A",
    "C4",
    "SIM",
    "TCH",
    "RUF",
]
ignore = ["RUF001", "RUF002", "RUF003"]

[tool.ruff.lint.per-file-ignores]
"testagent/gateway/*" = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["testagent"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    "appium.*",
    "celery.*",
    "jsonschema.*",
    "keyring.*",
    "kombu.*",
    "playwright.*",
    "selenium.*",
    "sentence_transformers.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["testagent.gateway.tasks"]
disable_error_code = ["untyped-decorator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    # Celery 任务协程在 asyncio.run mock 后不会被消费，属于正常测试行为
    "ignore:coroutine '.*_run' was never awaited:RuntimeWarning",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests requiring external services",
    "e2e: End-to-end tests",
    "docker: Docker-related tests requiring a running Docker daemon",
    "dual_backend: Tests that run against both SQLite and PostgreSQL backends",
]
